+++ src/api/fungible.rs
@@ -108,6 +108,8 @@ pub struct TransferApi {
 
     /// Optional change output: the rest of assets will be allocated here
     pub change: Option<OutPoint>,
+
+    ///
 }
 
 fn ticker_validator(name: &str) -> Result<(), String> {
